fix(tools/isobuild): :goal_net: abort build upon symlink in /public
When the bundler encounters a symbolic link in the /public folder, it currently just silently ignores the symlink and ALL SUBSEQUENT ASSETS.

This fix changes that behaviour to instead abort the build with an error message:

=> Errors prevented startup: While building the application: error: Symlink cycle detected at public/entities/fonts

This *may* potentially break builds which were previously silently ignoring symlinks (e.g. a symlink from /public/webfonts => /public/fonts ).